Make migration of expenses and expense types into a rake task
